+++ test/helpers/fixture_deploy_helper.rb
@@ -61,10 +61,11 @@ module FixtureDeployHelper
   end
 
   def deploy_dir_without_profiling(dir, wait: true, allow_protected_ns: false, prune: true, bindings: {},
-    sha: "k#{SecureRandom.hex(6)}", kubectl_instance: nil, max_watch_seconds: nil)
+    sha: "k#{SecureRandom.hex(6)}", kubectl_instance: nil, max_watch_seconds: nil, kubeconfig: nil)
     kubectl_instance ||= build_kubectl
 
     deploy = KubernetesDeploy::DeployTask.new(
+      kubeconfig: kubeconfig,
       namespace: @namespace,
       current_sha: sha,
       context: KubeclientHelper::TEST_CONTEXT,
@@ -127,8 +128,8 @@ module FixtureDeployHelper
     end
   end
 
-  def build_kubectl(log_failure_by_default: true, timeout: '5s')
+  def build_kubectl(kubeconfig: ENV["KUBECONFIG"], log_failure_by_default: true, timeout: '5s')
     KubernetesDeploy::Kubectl.new(namespace: @namespace, context: KubeclientHelper::TEST_CONTEXT, logger: logger,
-      log_failure_by_default: log_failure_by_default, default_timeout: timeout)
+      log_failure_by_default: log_failure_by_default, default_timeout: timeout, kubeconfig: kubeconfig)
   end
 end