+++ test/unit/kubernetes-deploy/deploy_task_test.rb
@@ -9,7 +9,7 @@ class DeployTaskTest < KubernetesDeploy::TestCase
   end
 
   def test_initializer
-    runner_with_env("/this-really-should/not-exist")
+    runner_with_kubeconfig("/this-really-should/not-exist")
     assert_logs_match("Configuration invalid")
     assert_logs_match("Kube config not found at /this-really-should/not-exist")
     assert_logs_match("Namespace must be specified")
@@ -19,18 +19,15 @@ class DeployTaskTest < KubernetesDeploy::TestCase
 
   private
 
-  def runner_with_env(value)
-    # TODO: Switch to --kubeconfig for kubectl shell out and pass env var as arg to DeployTask init
-    # Then fix this crappy env manipulation
-    with_env("KUBECONFIG", value) do
-      deploy = KubernetesDeploy::DeployTask.new(
-        namespace: "",
-        context: "",
-        logger: logger,
-        current_sha: "",
-        template_dir: "unknown",
-      )
-      deploy.run
-    end
+  def runner_with_kubeconfig(value)
+    deploy = KubernetesDeploy::DeployTask.new(
+      kubeconfig: value,
+      namespace: "",
+      context: "",
+      logger: logger,
+      current_sha: "",
+      template_dir: "unknown",
+    )
+    deploy.run
   end
 end