+++ test/unit/kubernetes-deploy/kubectl_test.rb
@@ -22,8 +22,11 @@ class KubectlTest < KubernetesDeploy::TestCase
   end
 
   def test_run_constructs_the_expected_command_and_returns_the_correct_values
-    stub_open3(%W(kubectl get pods -a --output=json --namespace=testn --context=testc --request-timeout=#{timeout}),
-      resp: "{ items: [] }")
+    stub_open3(
+      %W(kubectl get pods -a --output=json --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
+      resp: "{ items: [] }"
+    )
 
     out, err, st = build_kubectl.run("get", "pods", "-a", "--output=json")
     assert(st.success?)
@@ -32,61 +35,78 @@ class KubectlTest < KubernetesDeploy::TestCase
   end
 
   def test_run_omits_context_flag_if_use_context_is_false
-    stub_open3(%W(kubectl get pods -a --output=json --namespace=testn --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods -a --output=json --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --request-timeout=#{timeout}),
       resp: "{ items: [] }")
     build_kubectl.run("get", "pods", "-a", "--output=json", use_context: false)
   end
 
   def test_run_omits_namespace_flag_if_use_namespace_is_false
-    stub_open3(%W(kubectl get pods -a --output=json --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods -a --output=json --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--context=testc --request-timeout=#{timeout}),
       resp: "{ items: [] }")
     build_kubectl.run("get", "pods", "-a", "--output=json", use_namespace: false)
   end
 
   def test_run_logs_failures_when_log_failure_by_default_is_true_and_override_is_unspecified
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: true).run("get", "pods")
     assert_logs_match("[WARN]", 2)
   end
 
   def test_run_logs_failures_when_log_failure_by_default_is_true_and_override_is_also_true
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: true).run("get", "pods", log_failure: true)
     assert_logs_match("[WARN]", 2)
   end
 
   def test_run_does_not_log_failures_when_log_failure_by_default_is_true_and_override_is_false
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: true).run("get", "pods", log_failure: false)
     refute_logs_match("[WARN]")
   end
 
   def test_run_does_not_log_failures_when_log_failure_by_default_is_false_and_override_is_unspecified
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: false).run("get", "pods")
     refute_logs_match("[WARN]")
   end
 
   def test_run_does_not_log_failures_when_log_failure_by_default_is_false_and_override_is_also_false
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: false).run("get", "pods", log_failure: false)
     refute_logs_match("[WARN]")
   end
 
   def test_run_logs_failures_when_log_failure_by_default_is_false_and_override_is_true
-    stub_open3(%W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: "oops", success: false)
     build_kubectl(log_failure_by_default: false).run("get", "pods", log_failure: true)
     assert_logs_match("[WARN]", 2)
   end
 
   def test_run_with_multiple_attempts_retries_and_emits_failure_metrics
-    command = %W(kubectl get pods --namespace=testn --context=testc --request-timeout=#{timeout})
+    command = %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout})
     Open3.expects(:capture3).with(*command).times(5).returns(["", "oops", stub(success?: false)])
     kubectl = build_kubectl
     kubectl.expects(:retry_delay).returns(0).times(4)
@@ -102,7 +122,8 @@ class KubectlTest < KubernetesDeploy::TestCase
   def test_custom_timeout_is_used
     custom_kubectl = KubernetesDeploy::Kubectl.new(namespace: 'testn', context: 'testc', logger: logger,
     log_failure_by_default: true, default_timeout: '5s')
-    stub_open3(%w(kubectl get pods --namespace=testn --context=testc --request-timeout=5s),
+    stub_open3(
+      %W(kubectl get pods --kubeconfig=#{ENV['KUBECONFIG']} --namespace=testn --context=testc --request-timeout=5s),
       resp: "", err: "oops", success: false)
     custom_kubectl.run("get", "pods", log_failure: true)
     assert_logs_match("[WARN]", 2)
@@ -146,7 +167,10 @@ class KubectlTest < KubernetesDeploy::TestCase
   end
 
   def test_version_info_raises_if_command_fails
-    stub_open3(%W(kubectl version --context=testc --request-timeout=#{timeout}), resp: '', err: 'bad', success: false)
+    stub_open3(
+      %W(kubectl version --kubeconfig=#{ENV['KUBECONFIG']} --context=testc --request-timeout=#{timeout}),
+      resp: '', err: 'bad', success: false
+    )
     assert_raises_message(KubernetesDeploy::KubectlError, "Could not retrieve kubectl version info") do
       build_kubectl.version_info
     end
@@ -154,7 +178,9 @@ class KubectlTest < KubernetesDeploy::TestCase
 
   def test_run_with_raise_if_not_found_raises_the_correct_thing
     err = 'Error from server (NotFound): pods "foobar" not found'
-    stub_open3(%W(kubectl get pod foobar --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pod foobar --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: err, success: false)
     assert_raises_message(KubernetesDeploy::Kubectl::ResourceNotFoundError, err) do
       build_kubectl.run("get", "pod", "foobar", raise_if_not_found: true)
@@ -163,7 +189,9 @@ class KubectlTest < KubernetesDeploy::TestCase
 
   def test_run_with_raise_if_not_found_does_not_raise_on_other_errors
     err = 'Error from server (TooManyRequests): Please try again later'
-    stub_open3(%W(kubectl get pod foobar --namespace=testn --context=testc --request-timeout=#{timeout}),
+    stub_open3(
+      %W(kubectl get pod foobar --kubeconfig=#{ENV['KUBECONFIG']}) +
+      %W(--namespace=testn --context=testc --request-timeout=#{timeout}),
       resp: "", err: err, success: false)
     build_kubectl.run("get", "pod", "foobar", raise_if_not_found: true)
   end
@@ -175,7 +203,7 @@ class KubectlTest < KubernetesDeploy::TestCase
   end
 
   def stub_version_request(client:, server:)
-    stub_open3(%W(kubectl version --context=testc --request-timeout=#{timeout}), resp:
+    stub_open3(%W(kubectl version --kubeconfig=#{ENV['KUBECONFIG']} --context=testc --request-timeout=#{timeout}), resp:
       <<~STRING
         Client Version: #{client}
         Server Version: #{server}